Try int64

+++ arithmetic_lm/tokenizer.py
@@ -40,7 +40,7 @@ class CharTokenizer(Tokenizer):
     ) -> list[int] | torch.Tensor:
         tokens = [self.stoi_array[ord(char)] for char in text]
         if return_tensors:
-            return torch.tensor(tokens).int()
+            return torch.tensor(tokens, dtype=torch.int64)
         return tokens
 
     def decode(self, tokens: int | list[int] | torch.Tensor) -> str:
